requirements.txt
asgiref==3.7.2
boto3==1.28.17
botocore==1.31.17
defusedxml==0.7.1
diff-match-patch==20230430
Django==4.2.1
django-admin==2.0.2
django-cors-headers==4.0.0
django-environ==0.10.0
django-excel-response2==3.0.6
django-import-export==3.3.6
django-phonenumber-field==7.2.0
django-six==1.0.5
django-storages==1.13.2
djangorestframework==3.14.0
et-xmlfile==1.1.0
excel-base==1.0.4
gunicorn==20.1.0
isoweek==1.3.3
jmespath==1.0.1
MarkupPy==1.14
odfpy==1.4.1
openpyxl==3.1.2
Pillow==9.5.0
psycopg2==2.9.6
python-dateutil==2.8.2
pytz==2023.3
PyYAML==6.0.1
s3transfer==0.6.1
screen==1.0.1
six==1.16.0
sqlparse==0.4.4
tablib==3.5.0
TimeConvert==3.0.13
typing-extensions==4.6.3
tzlocal==5.2
urllib3==1.26.16
whitenoise==6.5.0
xlrd==2.0.1
xlwt==1.3.0
